Refactor code
Add new VariableExpression class which represents expressions
that are results of array or simple variable access. Class
VariableExpressionVisitor was also added. This class retrieves
the value from underlying variables. It also has an option
to return the pointer to the underlying value. This is used
in assignment expression to store a value to the corresponsign
memory.