Add default ret instruction to each function end
This handles a potential bug where the user specifies
if/else code block at the end of a non-void function
with both if and else containing a return. In such
a case, the compiler would generate a label to which
both of these blocks jump when finished. In this scenario,
the end block would never be reached. However, LLVM wouldn't
compile this code, since the end label wouldn't contain ret
or br instruction. This would be faulty behavior as the user
program would be correct. Default ret instruction ensures
the correctness of the above case.